requirements.txt
flask-restx~=0.5.1
Flask-SQLAlchemy~=2.5.1
Flask==2.1.2
SQLAlchemy~=1.4.40
marshmallow~=3.17.0
Werkzeug==2.1.2
pyjwt~=2.4.0
